third_party/blink/web_tests/external/wpt/css/css-rhythm/parsing/block-step-size-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property block-step-size value '0px'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-step-size value 'none'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-step-size value '100px'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-step-size value '2em'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-step-size value 'calc(10px + 0.5em)'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-step-size value 'calc(10px - 0.5em)'
  assert_true: block-step-size doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
